// src/enums/RouteNames.enum.ts
enum RouteNames {
  AUTH = 'auth',
  AUTH_ACCOUNTS = 'auth-accounts',
  AUTH_USERS = 'auth-users',
  LOGIN = 'login',
  LOGOUT = 'logout',
  MAIL = 'mail',
  MAIL_LOGS = 'mail-logs',
  MAIL_TEMPLATES = 'mail-templates',
  MAIL_SENT = 'mail-sent',
  MAIL_SETTINGS = 'mail-settings',
  SIGN_UP = 'sign-up',
  SIGN_UP_VERIFICATION = 'sign-up-verification',
  SIGN_UP_PAYMENT = 'sign-up-payment',
  SIGN_UP_PLAN = 'sign-up-plan',
  SIGN_UP_WELCOME = 'sign-up-welcome',
  SIGN_UP_PLEASE_VERIFY = 'sign-up-please-verify',
  PASSWORD_RESET_REQUEST = 'password-reset-request',
  PASSWORD_RESET = 'password-reset',
  TWO_STEP_AUTHENTICATION = 'two-step-authentication',
  TWO_STEP_AUTHENTICATION_CONFIRMATION = 'two-step-authentication-confirm',
  HUBS = 'hubs',
  ROOT = '/'
}

export default RouteNames